+++ hs-abci-example/package.yaml
@@ -55,6 +55,7 @@ dependencies:
 - hs-abci-server
 - hs-abci-types
 - hs-abci-extra
+- hs-abci-sdk
 - http-types
 - kan-extensions
 - katip
@@ -80,26 +81,11 @@ custom-setup:
 library:
   source-dirs:
   - src
-  - app
   ghc-options:
   - -fplugin=Polysemy.Plugin
   - -Werror
   - -Wall
   exposed-modules:
-  - Tendermint.SDK.Store
-  - Tendermint.SDK.Codec
-  - Tendermint.SDK.StoreQueries
-  - Tendermint.SDK.Subscription
-  - Tendermint.SDK.AuthTreeStore
-  - Tendermint.SDK.Router.Types
-  - Tendermint.SDK.Router.Class
-  - Tendermint.SDK.Router
-  - Tendermint.SDK.BaseApp
-  - Tendermint.SDK.Logger
-  - Tendermint.SDK.Logger.Katip
-  - Tendermint.SDK.Application
-  - Tendermint.SDK.Events
-
   - SimpleStorage.Server
   - SimpleStorage.Application
   - SimpleStorage.Handlers
@@ -142,29 +128,3 @@ tests:
       - hs-tendermint-client
       - hspec
       - QuickCheck
-
-  hs-abci-example-test:
-    main:                Spec.hs
-    source-dirs:         sdk-test
-    other-modules:
-    - Tendermint.SDK.Test.AuthTreeStoreSpec
-    # - Tendermint.SDK.Test.StoreExampleSpec
-    - Tendermint.SDK.Test.ModuleSpec
-
-    ghc-options:
-    - -Werror
-    - -Wall
-    - -threaded
-    - -rtsopts
-    - -with-rtsopts=-N
-
-    dependencies:
-      - generic-arbitrary
-      - hs-abci-example
-      - hs-abci-server
-      - hspec
-      - hspec-core
-      - hspec-discover
-      - binary
-      - QuickCheck
-      - quickcheck-instances
