--- hs-abci-example/sdk-test/Spec.hs
@@ -1,2 +1,0 @@
-{-# OPTIONS_GHC -F -pgmF hspec-discover #-}
-{-# OPTIONS_GHC -fno-warn-missing-import-lists #-}
