+++ hs-abci-example/src/SimpleStorage/Application.hs
