+++ hs-abci-example/src/SimpleStorage/Handlers.hs
