+++ hs-abci-example/src/SimpleStorage/Modules/SimpleStorage.hs
