+++ hs-abci-example/src/SimpleStorage/Server.hs
