+++ hs-abci-example/src/SimpleStorage/Types.hs
