+++ hs-abci-sdk/Setup.hs
@@ -1,2 +1,2 @@
-import           Data.ProtoLens.Setup
-main = defaultMainGeneratingProtos "protos"
+import           Distribution.Simple
+main = defaultMain
