--- hs-abci-sdk/docker-compose.yaml
@@ -1,26 +1,0 @@
-version: '3.7'
-services:
-  tendermint-init:
-    image: tendermint/tendermint:v0.32.2
-    command: init
-    volumes:
-        - /tmp:/tendermint
-  tendermint:
-    depends_on:
-      - tendermint-init
-      - simple_storage
-    image: tendermint/tendermint:v0.32.2
-    command: node --proxy_app=tcp://simple_storage:26658
-    volumes:
-      - /tmp:/tendermint
-    restart: always
-    ports:
-      - "26656-26657:26656-26657"
-  simple_storage:
-    build:
-      context: ../.
-      dockerfile: hs-abci-example/Dockerfile
-    restart: always
-    expose:
-      - "26658"
-
