+++ hs-abci-sdk/src/Tendermint/SDK/Logger/Katip.hs
