+++ stack.yaml
@@ -7,6 +7,7 @@ packages:
 - ./hs-abci-server
 - ./hs-abci-extra
 - ./hs-abci-example
+- ./hs-abci-sdk
 
 extra-deps:
   - proto-lens-runtime-0.5.0.0
